requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
pandas==2.0.3
sqlite3
requests==2.31.0
